Update AssetLintingSettingsProvider class
- Remove trailing spaces
- Reformat code
- Move to new namespace
- Update using references
