+++ Editor/Settings/AssetLintingSettingsProvider.cs.meta
